refactor: rename config folder to configs
