+++ src/main/java/com/cewar/configs/GameWebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.cewar.config;
+package com.cewar.configs;
 // package com.cewar.game;
 
 // import org.springframework.context.annotation.Configuration;
