+++ src/main/java/com/cewar/configs/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.cewar.config;
+package com.cewar.configs;
 
 import static org.springframework.security.config.Customizer.withDefaults;
